// frontend/sdk/models/ScoreboardRoundDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Unpaused Spotify API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { ScoreboardPlayerRoundDto } from './ScoreboardPlayerRoundDto';
import {
    ScoreboardPlayerRoundDtoFromJSON,
    ScoreboardPlayerRoundDtoFromJSONTyped,
    ScoreboardPlayerRoundDtoToJSON,
    ScoreboardPlayerRoundDtoToJSONTyped,
} from './ScoreboardPlayerRoundDto';

/**
 * 
 * @export
 * @interface ScoreboardRoundDto
 */
export interface ScoreboardRoundDto {
    /**
     * Round index (0-based)
     * @type {number}
     * @memberof ScoreboardRoundDto
     */
    roundIndex: number;
    /**
     * 
     * @type {Array<ScoreboardPlayerRoundDto>}
     * @memberof ScoreboardRoundDto
     */
    players: Array<ScoreboardPlayerRoundDto>;
    /**
     * Track name (after round complete)
     * @type {string}
     * @memberof ScoreboardRoundDto
     */
    trackName?: string;
    /**
     * Artist name (after round complete)
     * @type {string}
     * @memberof ScoreboardRoundDto
     */
    artistName?: string;
    /**
     * Album image URL
     * @type {string}
     * @memberof ScoreboardRoundDto
     */
    albumImageUrl?: string;
}

/**
 * Check if a given object implements the ScoreboardRoundDto interface.
 */
export function instanceOfScoreboardRoundDto(value: object): value is ScoreboardRoundDto {
    if (!('roundIndex' in value) || value['roundIndex'] === undefined) return false;
    if (!('players' in value) || value['players'] === undefined) return false;
    return true;
}

export function ScoreboardRoundDtoFromJSON(json: any): ScoreboardRoundDto {
    return ScoreboardRoundDtoFromJSONTyped(json, false);
}

export function ScoreboardRoundDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): ScoreboardRoundDto {
    if (json == null) {
        return json;
    }
    return {
        
        'roundIndex': json['roundIndex'],
        'players': ((json['players'] as Array<any>).map(ScoreboardPlayerRoundDtoFromJSON)),
        'trackName': json['trackName'] == null ? undefined : json['trackName'],
        'artistName': json['artistName'] == null ? undefined : json['artistName'],
        'albumImageUrl': json['albumImageUrl'] == null ? undefined : json['albumImageUrl'],
    };
}

export function ScoreboardRoundDtoToJSON(json: any): ScoreboardRoundDto {
    return ScoreboardRoundDtoToJSONTyped(json, false);
}

export function ScoreboardRoundDtoToJSONTyped(value?: ScoreboardRoundDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'roundIndex': value['roundIndex'],
        'players': ((value['players'] as Array<any>).map(ScoreboardPlayerRoundDtoToJSON)),
        'trackName': value['trackName'],
        'artistName': value['artistName'],
        'albumImageUrl': value['albumImageUrl'],
    };
}